apps/processor/requirements.txt
upstash-redis
pymongo[srv]
python-dotenv
sentence-transformers
torch
certifi